+++ index.ts
@@ -1,12 +1,12 @@
 // Default export of D2Api
-import D2Api from "./api";
+import D2Api from "./src/api";
 export default D2Api;
 
 // Additional API exports
-export * from "./api";
+export * from "./src/api";
 
 // React exports
 
 // Type exports
-export * from "./types/FetchError";
-export * from "./types/JsonValue";
+export * from "./src/types/FetchError";
+export * from "./src/types/JsonValue";
