+++ src/types/FetchError.ts
